fix bug

+++ app/src/main/java/com/example/bat2/Message.java
@@ -1,24 +1,24 @@
 package com.example.bat2;
 
 public class Message {
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private String title;
     private String body;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
